requirements.txt
attrs==23.1.0
certifi==2023.7.22
exceptiongroup==1.1.2
h11==0.14.0
idna==3.4
iniconfig==2.0.0
outcome==1.2.0
packaging==23.1
Pillow==10.0.0
pluggy==1.2.0
PySocks==1.7.1
pytest==7.4.0
selenium==4.10.0
sniffio==1.3.0
sortedcontainers==2.4.0
trio==0.22.2
trio-websocket==0.10.3
urllib3==2.0.4
wsproto==1.2.0
